refactor: group imports in index files
