+++ src/data/test/mock-http-client.ts
@@ -1,12 +1,16 @@
-import { HttpPostClient, HttpPostParams } from '@/data/protocols/http/http-post-client'
-import { HttpResponse, HttpStatusCode } from '@/data/protocols/http/http-response'
+import {
+  HttpPostClient,
+  HttpPostParams,
+  HttpResponse,
+  HttpStatusCode
+} from '@/data/protocols/http'
 
 export class HttpPostClientSpy<T, R> implements HttpPostClient<T, R> {
-  url?: string
-  body?: T
+  url?: string;
+  body?: T;
   response: HttpResponse<R> = {
     statusCode: HttpStatusCode.ok
-  }
+  };
 
   async post (params: HttpPostParams<T>): Promise<HttpResponse<R>> {
     this.url = params.url
