+++ src/domain/test/mock-account.ts
@@ -1,6 +1,6 @@
-import { AuthenticationParams } from '@/domain/usecases/autentication'
 import faker from 'faker'
-import { AccountModel } from '../models/account-model'
+import { AuthenticationParams } from '@/domain/usecases'
+import { AccountModel } from '../models'
 
 export const mockAuthentication = (): AuthenticationParams => ({
   email: faker.internet.email(),
